fix(emit): guard async super optional calls
