+++ crates/tsz-emitter/src/emitter/declarations/mod.rs
@@ -1,6 +1,7 @@
 pub(in crate::emitter) mod class;
 mod class_members;
 mod class_members_async_generator;
+mod class_members_async_method;
 mod core;
 mod namespace;
 mod namespace_body;
